+++ src/com/game/object/item/RedFlower.java
@@ -24,7 +24,7 @@ public class RedFlower extends GameObject{
 	@Override
 	public void render(Graphics g) {
 //		showBounds(g);
-		animation.drawAnimation(g, (int) x, (int) y, (int) width, (int) height);
+		animation.drawAnimation(g, x, y, width, height);
 	}
 
 	@Override
